Siirrä validaatio DateValidation -luokkaan
